ISSUE #274 * Move schemas to an independent file.
