+++ minos/common/protocol/avro/__init__.py
@@ -1,0 +1,15 @@
+"""
+Copyright (C) 2021 Clariteia SL
+
+This file is part of minos framework.
+
+Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
+"""
+from .protocols import (
+    MinosAvroProtocol,
+    MinosAvroValuesDatabase,
+)
+from .schemas import (
+    DATABASE_AVRO_SCHEMA,
+    MESSAGE_AVRO_SCHEMA,
+)
